// types/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type CastCredit = Person & {
  __typename?: 'CastCredit';
  character: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  order: Scalars['Int']['output'];
  profilePath?: Maybe<Scalars['String']['output']>;
};

export type CrewCredit = Person & {
  __typename?: 'CrewCredit';
  id: Scalars['ID']['output'];
  jobs: Array<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  profilePath?: Maybe<Scalars['String']['output']>;
};

export type List = {
  __typename?: 'List';
  id: Scalars['ID']['output'];
  items: Array<MovieOrPerson>;
  name: Scalars['String']['output'];
  type: ListType;
};

export enum ListType {
  Movie = 'MOVIE',
  Person = 'PERSON'
}

export type Movie = {
  __typename?: 'Movie';
  id: Scalars['ID']['output'];
  imdbId: Scalars['ID']['output'];
  inWatchlist?: Maybe<Scalars['Boolean']['output']>;
  overview: Scalars['String']['output'];
  posterPath?: Maybe<Scalars['String']['output']>;
  providers: Array<Provider>;
  releaseDate?: Maybe<Scalars['String']['output']>;
  sentiment?: Maybe<Sentiment>;
  tagline: Scalars['String']['output'];
  title: Scalars['String']['output'];
  tomatometer?: Maybe<Tomatometer>;
  voteAverage: Scalars['Float']['output'];
  watched?: Maybe<Scalars['Boolean']['output']>;
};
// ...
export type MovieInfo = {
  __typename?: 'MovieInfo';
  id: Scalars['ID']['output'];
  imdbId: Scalars['ID']['output'];
  overview: Scalars['String']['output'];
  posterPath?: Maybe<Scalars['String']['output']>;
  releaseDate?: Maybe<Scalars['String']['output']>;
  tagline: Scalars['String']['output'];
  title: Scalars['String']['output'];
  tomatometer?: Maybe<Tomatometer>;
  voteAverage: Scalars['Float']['output'];
};
// ...
export type Person = {
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  profilePath?: Maybe<Scalars['String']['output']>;
};

export type PersonInfo = Person & {
  __typename?: 'PersonInfo';
  biography?: Maybe<Scalars['String']['output']>;
  favourited?: Maybe<Scalars['Boolean']['output']>;
  id: Scalars['ID']['output'];
  knownForDepartment?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  profilePath?: Maybe<Scalars['String']['output']>;
};

export type Provider = {
  __typename?: 'Provider';
  id: Scalars['ID']['output'];
  logo: Scalars['String']['output'];
  name: Scalars['String']['output'];
};
// ...
export enum Sentiment {
  Disliked = 'DISLIKED',
  Liked = 'LIKED',
  None = 'NONE'
}

export type Stats = {
  __typename?: 'Stats';
  favouritePeople: Scalars['Int']['output'];
  moviesLiked: Scalars['Int']['output'];
  watched: Scalars['Int']['output'];
  watchlist: Scalars['Int']['output'];
};

export type Tomatometer = {
  __typename?: 'Tomatometer';
  consensus?: Maybe<Scalars['String']['output']>;
  link: Scalars['String']['output'];
  score: Scalars['Int']['output'];
  state: TomatometerState;
};

export enum TomatometerState {
  CertifiedFresh = 'CERTIFIED_FRESH',
  Fresh = 'FRESH',
  Rotten = 'ROTTEN'
}